Cool updates to dashboard
