+++ dashboard/src-tauri/src/mcap_recorder.rs
@@ -97,9 +97,9 @@ impl McapRecorder {
     }
 
     pub fn append_vsr(&mut self, payload: &[u8]) -> Option<String> {
-        if self.writer.is_none() {
+        let Some(writer) = self.writer.as_mut() else {
             return None;
-        }
+        };
 
         let log_time = epoch_nanos();
         let header = MessageHeader {
@@ -109,11 +109,7 @@ impl McapRecorder {
             publish_time: log_time,
         };
 
-        let write_result = self
-            .writer
-            .as_mut()
-            .expect("checked writer.is_some()")
-            .write_to_known_channel(&header, payload);
+        let write_result = writer.write_to_known_channel(&header, payload);
         if let Err(err) = write_result {
             return Some(
                 self.disable_logging(format!("MCAP logging stopped: write failed: {err}")),
@@ -126,11 +122,11 @@ impl McapRecorder {
         if self.messages_since_flush >= MCAP_FLUSH_EVERY_MESSAGES
             || self.last_flush_at.elapsed() >= MCAP_FLUSH_INTERVAL
         {
-            let flush_result = self
-                .writer
-                .as_mut()
-                .expect("writer must still be available after successful write")
-                .flush();
+            let Some(writer) = self.writer.as_mut() else {
+                return None;
+            };
+
+            let flush_result = writer.flush();
             if let Err(err) = flush_result {
                 return Some(
                     self.disable_logging(format!("MCAP logging stopped: flush failed: {err}")),
